+++ lib/src/clixon_yang_internal.h
@@ -39,6 +39,16 @@
 #ifndef _CLIXON_YANG_INTERNAL_H_
 #define _CLIXON_YANG_INTERNAL_H_
 
+
+/*
+ * Actually cligen variable stuff XXX
+ */
+#define V_UNIQUE	0x01	/* Variable flag */
+#define V_UNSET		0x08	/* Variable is unset, ie no default */
+
+
+#define YANG_FLAG_MARK 0x01  /* Marker for dynamic algorithms, eg expand */
+
 /*! Yang type cache. Yang type statements can cache all typedef info here
  * @note unions not cached
 */
@@ -87,5 +97,21 @@ struct yang_stmt{
     yang_type_cache   *ys_typecache; /* If ys_keyword==Y_TYPE, cache all typedef data except unions */
     int               _ys_vector_i;   /* internal use: yn_each */
 };
+
+/* Yang data definition statement
+ * See RFC 7950 Sec 3:
+ *   o  data definition statement: A statement that defines new data
+ *      nodes.  One of "container", "leaf", "leaf-list", "list", "choice",
+ *      "case", "augment", "uses", "anydata", and "anyxml".
+ */
+#define yang_datadefinition(y) (yang_datanode(y) || (y)->ys_keyword == Y_CHOICE || (y)->ys_keyword == Y_CASE || (y)->ys_keyword == Y_AUGMENT || (y)->ys_keyword == Y_USES)
+
+/* Yang schema node .
+ * See RFC 7950 Sec 3:
+ *    o  schema node: A node in the schema tree.  One of action, container,
+ *       leaf, leaf-list, list, choice, case, rpc, input, output,
+ *       notification, anydata, and anyxml.
+ */
+#define yang_schemanode(y) (yang_datanode(y) || (y)->ys_keyword == Y_RPC || (y)->ys_keyword == Y_CHOICE || (y)->ys_keyword == Y_CASE || (y)->ys_keyword == Y_INPUT || (y)->ys_keyword == Y_OUTPUT || (y)->ys_keyword == Y_NOTIFICATION)
 
 #endif  /* _CLIXON_YANG_INTERNAL_H_ */