+++ lib/src/clixon_yang_type.c
@@ -67,6 +67,7 @@
 #include "clixon_plugin.h"
 #include "clixon_options.h"
 #include "clixon_yang.h"
+#include "clixon_yang_internal.h" /* internal */
 #include "clixon_yang_type.h"
 
 /*